feat: move agent middlewares into server
