+++ .deploy/api/package.json
@@ -28,7 +28,7 @@
     "start": "nx serve",
     "build": "nx build",
     "test": "nx test",
-    "build:plugins": "pnpm nx build agent-middlewares && pnpm nx build vlm-default && pnpm nx build draft"
+    "build:plugins": "pnpm nx build vlm-default && pnpm nx build draft"
   },
   "dependencies": {
     "@langchain/community": "0.3.53",
