+++ package.json
@@ -44,7 +44,7 @@
     "format:fix": "prettier --write --ignore-unknown",
     "prepare": "command -v git >/dev/null 2>&1 && git config core.hooksPath git-hooks || exit 0",
     "prebuild": "rimraf dist coverage",
-    "build:plugins": "pnpm nx build agent-middlewares && pnpm nx build vlm-default && pnpm nx build draft",
+    "build:plugins": "pnpm nx build vlm-default && pnpm nx build draft",
     "build:embed": "pnpm exec terser apps/cloud/src/assets/embed.js --compress --mangle --output apps/cloud/src/assets/embed.min.js && printf '\\n' >> apps/cloud/src/assets/embed.min.js",
     "build": "pnpm nx build store && pnpm nx build core && pnpm nx build echarts && pnpm nx build sql && pnpm nx build xmla && pnpm -C ./packages/duckdb build:release && pnpm nx build duckdb && pnpm b:p:copilot && pnpm b:p:angular",
     "b:p:duckdb": "pnpm -C ./packages/duckdb build:release && pnpm nx build duckdb",
