--- packages/plugins/agent-middlewares/CHANGELOG.md
@@ -1,165 +1,0 @@
-# @xpert-ai/plugin-agent-middlewares
-
-## 8.0.0
-
-### Patch Changes
-
-- Updated dependencies [6f679b8]
-  - @xpert-ai/plugin-sdk@3.14.0
-  - @xpert-ai/contracts@3.15.0
-
-## 7.0.0
-
-### Patch Changes
-
-- Updated dependencies [54cff15]
-- Updated dependencies [6978bfd]
-  - @xpert-ai/plugin-sdk@3.13.0
-  - @xpert-ai/contracts@3.14.0
-
-## 6.0.0
-
-### Patch Changes
-
-- Updated dependencies [e6528c8]
-- Updated dependencies [f23228b]
-  - @xpert-ai/contracts@3.13.0
-  - @xpert-ai/plugin-sdk@3.12.2
-
-## 5.0.1
-
-### Patch Changes
-
-- Updated dependencies [6a17eca]
-  - @xpert-ai/plugin-sdk@3.12.1
-  - @xpert-ai/contracts@3.12.1
-
-## 5.0.0
-
-### Patch Changes
-
-- Updated dependencies [d017897]
-  - @xpert-ai/plugin-sdk@3.12.0
-  - @xpert-ai/contracts@3.12.0
-
-## 4.0.1
-
-### Patch Changes
-
-- Updated dependencies [7418eef]
-  - @xpert-ai/plugin-sdk@3.11.2
-
-## 4.0.0
-
-### Patch Changes
-
-- Updated dependencies [d92d0f2]
-  - @xpert-ai/plugin-sdk@4.0.0
-  - @xpert-ai/contracts@3.11.0
-
-## 3.0.0
-
-### Patch Changes
-
-- Updated dependencies [49101da]
-  - @xpert-ai/plugin-sdk@3.11.0
-  - @xpert-ai/contracts@3.10.1
-
-## 2.0.0
-
-### Patch Changes
-
-- Updated dependencies [a83c9ea]
-  - @xpert-ai/plugin-sdk@3.10.0
-
-## 1.0.0
-
-### Patch Changes
-
-- Updated dependencies [df9d7e2]
-  - @xpert-ai/plugin-sdk@4.0.0
-  - @xpert-ai/contracts@3.10.0
-
-## 0.0.9
-
-### Patch Changes
-
-- Updated dependencies [2acc11a]
-  - @xpert-ai/contracts@3.9.9
-  - @xpert-ai/plugin-sdk@3.9.9
-
-## 0.0.8
-
-### Patch Changes
-
-- Updated dependencies [2558760]
-  - @xpert-ai/contracts@3.9.8
-  - @xpert-ai/plugin-sdk@3.9.8
-
-## 0.0.7
-
-### Patch Changes
-
-- Updated dependencies [9e37ff9]
-  - @xpert-ai/plugin-sdk@3.9.5
-  - @xpert-ai/contracts@3.9.5
-
-## 0.0.6
-
-### Patch Changes
-
-- Updated dependencies [07057a6]
-  - @xpert-ai/contracts@3.9.4
-  - @xpert-ai/plugin-sdk@3.9.4
-
-## 0.0.5
-
-### Patch Changes
-
-- Updated dependencies [ea234e5]
-- Updated dependencies [4920c48]
-  - @xpert-ai/plugin-sdk@3.9.3
-  - @xpert-ai/contracts@3.9.3
-
-## 0.0.4
-
-### Patch Changes
-
-- Updated dependencies [8187f99]
-  - @xpert-ai/plugin-sdk@3.9.2
-  - @xpert-ai/contracts@3.9.2
-
-## 0.0.3
-
-### Patch Changes
-
-- Updated dependencies [e040933]
-  - @xpert-ai/plugin-sdk@3.9.1
-  - @xpert-ai/contracts@3.9.1
-
-## 0.0.2
-
-### Patch Changes
-
-- Updated dependencies [4dcf5b5]
-- Updated dependencies [7fff870]
-- Updated dependencies [c76facd]
-- Updated dependencies [5b5c8ef]
-  - @xpert-ai/plugin-sdk@3.9.0
-  - @xpert-ai/contracts@3.9.0
-
-## 0.0.2-beta.1
-
-### Patch Changes
-
-- Updated dependencies
-  - @xpert-ai/plugin-sdk@3.9.0-beta.2
-  - @xpert-ai/contracts@3.9.0-beta.2
-
-## 0.0.2-beta.0
-
-### Patch Changes
-
-- Updated dependencies
-  - @xpert-ai/plugin-sdk@3.9.0-beta.1
-  - @xpert-ai/contracts@3.9.0-beta.1
