--- packages/plugins/agent-middlewares/README.md
@@ -1,11 +1,0 @@
-# agent-middlewares
-
-This library was generated with [Nx](https://nx.dev).
-
-## Building
-
-Run `nx build agent-middlewares` to build the library.
-
-## Running unit tests
-
-Run `nx test agent-middlewares` to execute the unit tests via [Jest](https://jestjs.io).
