--- packages/plugins/agent-middlewares/package.json
@@ -1,22 +1,0 @@
-{
-  "name": "@xpert-ai/plugin-agent-middlewares",
-  "version": "8.0.0",
-  "private": true,
-  "xpert": {
-    "plugin": {
-      "level": "system"
-    }
-  },
-  "dependencies": {
-    "node-pty": "^1.0.0"
-  },
-  "main": "./index.cjs",
-  "module": "./index.js",
-  "peerDependencies": {
-    "@xpert-ai/contracts": "workspace:*",
-    "@nestjs/common": ">=10",
-    "@nestjs/core": ">=10",
-    "@xpert-ai/plugin-sdk": "workspace:*",
-    "zod": ">=3.22.4"
-  }
-}
