--- packages/plugins/agent-middlewares/project.json
@@ -1,52 +1,0 @@
-{
-  "name": "agent-middlewares",
-  "$schema": "../../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "packages/plugins/agent-middlewares/src",
-  "projectType": "library",
-  "release": {
-    "version": {
-      "generatorOptions": {
-        "packageRoot": "dist/{projectRoot}",
-        "currentVersionResolver": "git-tag"
-      }
-    }
-  },
-  "tags": [],
-  "targets": {
-    "build": {
-      "executor": "@nx/rollup:rollup",
-      "outputs": ["{options.outputPath}"],
-      "options": {
-        "outputPath": "dist/packages/plugins/agent-middlewares",
-        "main": "packages/plugins/agent-middlewares/src/index.ts",
-        "tsConfig": "packages/plugins/agent-middlewares/tsconfig.lib.json",
-        "assets": [],
-        "project": "packages/plugins/agent-middlewares/package.json",
-        "compiler": "swc",
-        "format": ["cjs", "esm"],
-        "external": [
-          "@nestjs/common",
-          "@nestjs/core",
-          "@xpert-ai/plugin-sdk",
-          "@nestjs/microservices",
-          "kafkajs",
-          "amqplib",
-          "mqtt",
-          "nats"
-        ]
-      }
-    },
-    "nx-release-publish": {
-      "options": {
-        "packageRoot": "dist/{projectRoot}"
-      }
-    },
-    "test": {
-      "executor": "@nx/jest:jest",
-      "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
-      "options": {
-        "jestConfig": "packages/plugins/agent-middlewares/jest.config.ts"
-      }
-    }
-  }
-}
