--- packages/plugins/agent-middlewares/src/index.ts
@@ -1,38 +1,0 @@
-import { z } from 'zod'
-import type { XpertPlugin } from '@xpert-ai/plugin-sdk'
-import { AgentMiddlewaresModule } from './lib/middlewares.module'
-
-const ConfigSchema = z.object({})
-
-const plugin: XpertPlugin<z.infer<typeof ConfigSchema>> = {
-  meta: {
-    name: '@xpert-ai/plugin-agent-middlewares',
-    version: '0.0.1',
-    level: 'system',
-    category: 'middleware',
-    icon: {
-      type: 'svg',
-      value: `<svg viewBox="-2 0 60 60" version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" fill="currentColor"><g id="SVGRepo_bgCarrier" stroke-width="0"></g><g id="SVGRepo_tracerCarrier" stroke-linecap="round" stroke-linejoin="round"></g><g id="SVGRepo_iconCarrier"> <title>android-brain</title> <desc>Created with Sketch.</desc> <defs> </defs> <g id="People" stroke="none" stroke-width="1" fill="none" fill-rule="evenodd"> <g id="Icon-9" transform="translate(-2.000000, 0.000000)" fill="currentColor"> <path d="M22,27 C22,27.551 22.448,28 23,28 C23.552,28 24,27.551 24,27 C24,26.449 23.552,26 23,26 C22.448,26 22,26.449 22,27 L22,27 Z M24,45 C24,44.449 23.552,44 23,44 C22.448,44 22,44.449 22,45 C22,45.551 22.448,46 23,46 C23.552,46 24,45.551 24,45 L24,45 Z M16,27 C16,26.449 15.552,26 15,26 C14.448,26 14,26.449 14,27 C14,27.551 14.448,28 15,28 C15.552,28 16,27.551 16,27 L16,27 Z M16,39 C16,38.449 15.552,38 15,38 C14.448,38 14,38.449 14,39 C14,39.551 14.448,40 15,40 C15.552,40 16,39.551 16,39 L16,39 Z M22,17 C22,17.551 22.448,18 23,18 C23.552,18 24,17.551 24,17 C24,16.449 23.552,16 23,16 C22.448,16 22,16.449 22,17 L22,17 Z M16,18 L16,24.184 C16.848,24.486 17.514,25.153 17.815,26 L20.185,26 C20.599,24.838 21.698,24 23,24 C24.654,24 26,25.346 26,27 C26,28.302 25.161,29.401 24,29.816 L24,42.184 C25.161,42.599 26,43.698 26,45 C26,46.302 25.161,47.401 24,47.816 L24,51 C24,51.552 23.553,52 23,52 C22.447,52 22,51.552 22,51 L22,47.815 C20.979,47.452 20.218,46.558 20.047,45.46 L16.293,41.707 C16.29,41.705 16.29,41.701 16.287,41.698 C15.896,41.886 15.463,42 15,42 C13.346,42 12,40.654 12,39 C12,37.346 13.346,36 15,36 C16.654,36 18,37.346 18,39 C18,39.462 17.886,39.896 17.698,40.287 C17.701,40.29 17.704,40.291 17.707,40.293 L20.614,43.201 C20.965,42.737 21.443,42.383 22,42.184 L22,29.815 C21.152,29.514 20.486,28.847 20.185,28 L17.815,28 C17.401,29.162 16.302,30 15,30 C13.346,30 12,28.654 12,27 C12,25.698 12.839,24.599 14,24.184 L14,17 C14,16.448 14.447,16 15,16 L20.185,16 C20.486,15.153 21.152,14.486 22,14.184 L22,11 C22,10.448 22.447,10 23,10 C23.553,10 24,10.448 24,11 L24,14.184 C25.161,14.599 26,15.698 26,17 C26,18.654 24.654,20 23,20 C21.698,20 20.599,19.162 20.185,18 L16,18 Z M55.633,37.245 C57.122,35.936 58,34.042 58,32 C58,29.86 57.049,27.894 55.415,26.574 C55.8,25.771 56,24.896 56,24 C56,21.936 54.917,20.035 53.208,18.95 C53.724,18.051 54,17.037 54,16 C54,13.42 52.361,11.177 49.989,10.349 C49.996,10.232 50,10.116 50,10 C50,6.691 47.309,4 44,4 C43.193,4 42.407,4.164 41.673,4.48 C40.813,1.876 38.34,0 35.5,0 C31.292,0 29.421,3.796 29.02,5.804 C29.013,5.837 29.022,5.869 29.02,5.903 C29.017,5.937 29,5.965 29,6 L29,52.94 C28.774,55.779 26.363,58 23.5,58 C20.647,58 18.243,55.771 18.026,52.924 C17.986,52.403 17.522,52 17,52 C16.688,52 16.394,52.146 16.204,52.394 C15.428,53.415 14.26,54 13,54 C10.794,54 9,52.206 9,50 C9,49.092 9.316,48.224 9.916,47.489 C10.186,47.158 10.216,46.691 9.989,46.328 C9.763,45.965 9.335,45.785 8.913,45.884 C8.578,45.962 8.279,46 8,46 C5.794,46 4,44.206 4,42 C4,40.351 5.047,38.851 6.604,38.267 C6.975,38.128 7.229,37.784 7.252,37.39 C7.275,36.995 7.063,36.623 6.713,36.442 C5.039,35.579 4,33.876 4,32 C4,30.236 4.907,28.636 6.427,27.72 C6.664,27.577 6.831,27.342 6.889,27.071 C6.946,26.799 6.888,26.517 6.729,26.29 C6.252,25.609 6,24.818 6,24 C6,22.339 7.058,20.836 8.631,20.259 C8.942,20.145 9.177,19.884 9.258,19.563 C9.338,19.241 9.253,18.9 9.032,18.653 C8.366,17.91 8,16.968 8,16 C8,14.062 9.389,12.411 11.301,12.073 C11.569,12.025 11.808,11.87 11.959,11.643 C12.11,11.416 12.162,11.137 12.104,10.871 C12.034,10.559 12,10.274 12,10 C12,7.794 13.794,6 16,6 C16.866,6 17.701,6.289 18.416,6.836 C18.707,7.058 19.096,7.104 19.429,6.956 C19.764,6.807 19.99,6.488 20.021,6.124 C20.211,3.812 22.179,2 24.5,2 C25.614,2 26.535,2.358 27.315,3.094 C27.718,3.473 28.351,3.454 28.729,3.052 C29.108,2.65 29.09,2.018 28.688,1.639 C27.535,0.551 26.126,0 24.5,0 C21.66,0 19.187,1.876 18.327,4.48 C17.593,4.164 16.807,4 16,4 C12.691,4 10,6.691 10,10 C10,10.116 10.004,10.232 10.011,10.349 C7.639,11.177 6,13.42 6,16 C6,17.037 6.276,18.051 6.792,18.95 C5.083,20.035 4,21.936 4,24 C4,24.896 4.2,25.771 4.585,26.574 C2.951,27.894 2,29.86 2,32 C2,34.042 2.878,35.936 4.367,37.245 C2.906,38.362 2,40.116 2,42 C2,45.093 4.353,47.647 7.363,47.966 C7.123,48.614 7,49.298 7,50 C7,53.309 9.691,56 13,56 C14.242,56 15.424,55.623 16.416,54.932 C17.436,57.886 20.255,60 23.5,60 C27.408,60 30.701,56.958 30.997,53.076 C30.998,53.063 30.992,53.051 30.992,53.038 C30.993,53.024 31,53.014 31,53 L31,6.11 C31.148,5.496 32.135,2 35.5,2 C37.821,2 39.789,3.812 39.979,6.124 C40.01,6.488 40.236,6.807 40.571,6.956 C40.904,7.104 41.293,7.058 41.584,6.836 C42.299,6.289 43.134,6 44,6 C46.206,6 48,7.794 48,10 C48,10.274 47.966,10.559 47.896,10.871 C47.838,11.137 47.89,11.416 48.041,11.643 C48.192,11.87 48.431,12.025 48.699,12.073 C50.611,12.411 52,14.062 52,16 C52,16.968 51.634,17.91 50.968,18.653 C50.747,18.9 50.662,19.241 50.742,19.563 C50.823,19.884 51.058,20.145 51.369,20.259 C52.942,20.836 54,22.339 54,24 C54,24.818 53.748,25.609 53.271,26.29 C53.112,26.517 53.054,26.799 53.111,27.071 C53.169,27.342 53.336,27.577 53.573,27.72 C55.093,28.636 56,30.236 56,32 C56,33.876 54.961,35.579 53.287,36.442 C52.937,36.623 52.725,36.995 52.748,37.39 C52.771,37.784 53.025,38.128 53.396,38.267 C54.953,38.851 56,40.351 56,42 C56,44.206 54.206,46 52,46 C51.721,46 51.422,45.962 51.087,45.884 C50.667,45.785 50.238,45.965 50.011,46.328 C49.784,46.691 49.814,47.158 50.084,47.489 C50.684,48.224 51,49.092 51,50 C51,52.206 49.206,54 47,54 C45.736,54 44.57,53.408 43.802,52.375 C43.544,52.03 43.095,51.889 42.686,52.023 C42.538,52.072 42.408,52.153 42.303,52.256 C42.117,52.422 41.994,52.658 41.974,52.924 C41.757,55.771 39.353,58 36.5,58 C35.243,58 34.057,57.587 33.07,56.806 C32.635,56.463 32.008,56.537 31.665,56.969 C31.322,57.402 31.396,58.031 31.828,58.374 C33.153,59.422 34.813,60 36.5,60 C39.747,60 42.568,57.882 43.586,54.925 C44.575,55.621 45.755,56 47,56 C50.309,56 53,53.309 53,50 C53,49.298 52.877,48.614 52.637,47.966 C55.647,47.647 58,45.093 58,42 C58,40.116 57.094,38.362 55.633,37.245 L55.633,37.245 Z M38,45 C38,45.552 38.447,46 39,46 C40.458,46 42,47.542 42,49 C42,49.552 42.447,50 43,50 C43.553,50 44,49.552 44,49 C44,46.43 41.57,44 39,44 C38.447,44 38,44.448 38,45 L38,45 Z M47,18 C45.542,18 44,16.458 44,15 C44,14.448 43.553,14 43,14 C42.447,14 42,14.448 42,15 C42,17.57 44.43,20 47,20 C47.553,20 48,19.552 48,19 C48,18.448 47.553,18 47,18 L47,18 Z M35,40 C35.553,40 36,39.552 36,39 C36,37.542 37.542,36 39,36 C39.553,36 40,35.552 40,35 C40,34.448 39.553,34 39,34 C36.43,34 34,36.43 34,39 C34,39.552 34.447,40 35,40 L35,40 Z M39,28 C39.553,28 40,27.552 40,27 C40,26.448 39.553,26 39,26 C37.542,26 36,24.458 36,23 C36,22.448 35.553,22 35,22 C34.447,22 34,22.448 34,23 C34,25.57 36.43,28 39,28 L39,28 Z M49,24 C48.447,24 48,24.448 48,25 C48,26.458 46.458,28 45,28 C44.447,28 44,28.448 44,29 C44,29.552 44.447,30 45,30 C47.57,30 50,27.57 50,25 C50,24.448 49.553,24 49,24 L49,24 Z M45,36 C44.447,36 44,36.448 44,37 C44,37.552 44.447,38 45,38 C46.458,38 48,39.542 48,41 C48,41.552 48.447,42 49,42 C49.553,42 50,41.552 50,41 C50,38.43 47.57,36 45,36 L45,36 Z M34,14 C34,11.43 36.43,9 39,9 C39.553,9 40,9.448 40,10 C40,10.552 39.553,11 39,11 C37.542,11 36,12.542 36,14 C36,14.552 35.553,15 35,15 C34.447,15 34,14.552 34,14 L34,14 Z" id="android-brain"> </path> </g> </g> </g></svg>`,
-      color: '#00d2e6'
-    },
-    displayName: 'Built-in Agent Middlewares',
-    description: 'Provide Built-in Agent Middlewares for XpertAI Agents',
-    keywords: ['middleware', 'agent'],
-    author: 'XpertAI Team'
-  },
-  config: {
-    schema: ConfigSchema
-  },
-  register(ctx) {
-    ctx.logger.log('register agent-middlewares plugin')
-    return { module: AgentMiddlewaresModule, global: true }
-  },
-  async onStart(ctx) {
-    ctx.logger.log('agent-middlewares plugin started')
-  },
-  async onStop(ctx) {
-    ctx.logger.log('agent-middlewares plugin stopped')
-  }
-}
-
-export default plugin
