+++ packages/server-ai/src/sandbox/middlewares/index.ts
@@ -1,2 +1,3 @@
 export * from './sandbox-shell.middleware'
 export * from './sandbox-service.middleware'
+export * from './sandbox-file.middleware'
