+++ packages/server-ai/src/xpert-middleware/agent-middleware-registry.scope.spec.ts
@@ -1,7 +1,12 @@
 import 'reflect-metadata'
 import { Reflector } from '@nestjs/core'
 import { AgentMiddlewareRegistry, RequestContext, setDefaultTenantId } from '@xpert-ai/plugin-sdk'
+import { ClientEffectMiddleware } from './client-effect.middleware'
 import { HumanInTheLoopMiddleware } from './human-in-the-loop.middleware'
+import { LLMToolSelectorNameMiddleware } from './llm-tool-selector.middleware'
+import { StructuredOutputMiddleware } from './structured-output.middleware'
+import { SummarizationMiddleware } from './summarization.middleware'
+import { TodoListMiddleware } from './todo-list.middleware'
 
 describe('built-in agent middleware registry scope', () => {
     beforeEach(() => {
@@ -13,9 +18,18 @@ describe('built-in agent middleware registry scope', () => {
         setDefaultTenantId('tenant-default')
 
         const registry = new AgentMiddlewareRegistry({} as any, new Reflector())
-        const builtinMiddleware = new HumanInTheLoopMiddleware()
+        const builtinMiddlewares = [
+            new HumanInTheLoopMiddleware(),
+            new ClientEffectMiddleware(),
+            new LLMToolSelectorNameMiddleware(),
+            new StructuredOutputMiddleware(),
+            new SummarizationMiddleware(),
+            new TodoListMiddleware()
+        ]
 
-        registry.upsert(builtinMiddleware)
+        for (const middleware of builtinMiddlewares) {
+            registry.upsert(middleware)
+        }
 
         jest.spyOn(RequestContext, 'getScope').mockReturnValue({
             tenantId: 'tenant-other',
@@ -25,7 +39,9 @@ describe('built-in agent middleware registry scope', () => {
         jest.spyOn(RequestContext, 'currentTenantId').mockReturnValue('tenant-other')
         jest.spyOn(RequestContext, 'getOrganizationId').mockReturnValue('org-other')
 
-        expect(registry.get(builtinMiddleware.meta.name, 'org-other')).toBe(builtinMiddleware)
-        expect(registry.list('org-other')).toContain(builtinMiddleware)
+        for (const middleware of builtinMiddlewares) {
+            expect(registry.get(middleware.meta.name, 'org-other')).toBe(middleware)
+            expect(registry.list('org-other')).toContain(middleware)
+        }
     })
 })
