+++ packages/server-ai/src/xpert-middleware/index.ts
@@ -1,7 +1,12 @@
 export * from './browser-automation.middleware'
+export * from './client-effect.middleware'
 export * from './client-tool.middleware'
 export * from './context-compression.middleware'
 export * from './human-in-the-loop.middleware'
+export * from './llm-tool-selector.middleware'
 export * from './office-automation.middleware'
 export * from './ralph-loop.middleware'
+export * from './structured-output.middleware'
+export * from './summarization.middleware'
+export * from './todo-list.middleware'
 export * from './xpert-middleware.module'
